refactor(letters): extract buildLetterTemplateData util for reuse
Moves contract+company+settings fetch and LetterTemplateData assembly
out of LetterDispatchDialog into a standalone util so BulkPrintDialog
can call it in a loop without duplicating the money arithmetic.
